wip: #60 moved requireLayoutProps to utils file in lib folder
