+++ apps/issues/pages/organizations/add.tsx
@@ -7,7 +7,7 @@ import { z } from 'zod';
 import FormInput from '../../components/FormInput';
 import Head from '../../components/Head';
 import { trpc } from '../../lib/trpc';
-import requireLayoutProps from '../../utils/requireLayoutProps';
+import { requireLayoutProps } from '../../lib/utils';
 
 const FormSchema = OrganizationModel.omit({ id: true });
 type FormSchemaType = z.infer<typeof FormSchema>;
