+++ apps/issues/pages/tickets/add.tsx
@@ -9,7 +9,7 @@ import { z } from 'zod';
 import FormInput from '../../components/FormInput';
 import Head from '../../components/Head';
 import { trpc } from '../../lib/trpc';
-import requireLayoutProps from '../../utils/requireLayoutProps';
+import { requireLayoutProps } from '../../lib/utils';
 
 interface QParams extends ParsedUrlQuery {
   project_id: string | undefined;
