+++ apps/issues/pages/tickets/settings.tsx
@@ -3,7 +3,7 @@ import { PencilIcon, TrashIcon } from '@heroicons/react/outline';
 import Head from '../../components/Head';
 import { TicketSettingsForm } from '../../components/TicketSettingsForm';
 import { trpc } from '../../lib/trpc';
-import requireLayoutProps from '../../utils/requireLayoutProps';
+import { requireLayoutProps } from '../../lib/utils';
 
 function Settings() {
   const { data: ticketPriorities } = trpc.ticket.priority.getAll.useQuery();
